Extract History Period Store
