+++ src/ServicePulse.Host/vue/src/components/monitoring/EndpointDetails.vue
@@ -18,6 +18,7 @@ import EndpointWorkload from "./EndpointWorkload.vue";
 import EndpointTimings from "./EndpointTimings.vue";
 import EndpointInstances from "./EndpointInstances.vue";
 import EndpointMessageTypes from "./EndpointMessageTypes.vue";
+import { useMonitoringHistoryPeriodStore } from "@/stores/MonitoringHistoryPeriodStore";
 
 const route = useRoute();
 const router = useRouter();
@@ -25,6 +26,7 @@ const endpointName = route.params.endpointName;
 let refreshInterval = undefined;
 
 const monitoringStore = useMonitoringStore();
+const monitoringHistoryPeriodStore = useMonitoringHistoryPeriodStore();
 
 const showInstancesBreakdown = ref(route?.query?.tab === "instancesBreakdown");
 
@@ -32,7 +34,8 @@ const loadedSuccessfully = ref(false);
 
 const endpoint = ref({});
 
-const { historyPeriod, negativeCriticalTimeIsPresent } = storeToRefs(monitoringStore);
+const { historyPeriod } = storeToRefs(monitoringHistoryPeriodStore);
+const { negativeCriticalTimeIsPresent } = storeToRefs(monitoringStore);
 
 watch(historyPeriod, (newValue) => {
   changeRefreshInterval(newValue.refreshIntervalVal);
