+++ src/ServicePulse.Host/vue/src/components/monitoring/MonitoringFilter.vue
@@ -4,12 +4,15 @@ import GroupBy from "./MonitoringGroupBy.vue";
 import PeriodSelector from "./MonitoringHistoryPeriod.vue";
 import { useMonitoringStore } from "@/stores/MonitoringStore";
 import { useRoute } from "vue-router";
+import { useMonitoringHistoryPeriodStore } from "@/stores/MonitoringHistoryPeriodStore";
 
 const monitoringStore = useMonitoringStore();
+const monitoringHistoryPeriodStore = useMonitoringHistoryPeriodStore();
+
 const filterString = ref(monitoringStore.filterString);
 const route = useRoute();
 
-watch(route, () => monitoringStore.setHistoryPeriod(route.params.historyPeriod), { deep: true, immediate: true, flush: "pre" });
+watch(route, () => monitoringHistoryPeriodStore.setHistoryPeriod(route.params.historyPeriod), { deep: true, immediate: true, flush: "pre" });
 watch(filterString, async (newValue) => {
   await monitoringStore.updateFilterString(newValue);
   filterString.value = monitoringStore.filterString;
