+++ src/ServicePulse.Host/vue/src/stores/MonitoringStore.js
@@ -2,30 +2,15 @@ import { defineStore, acceptHMRUpdate } from "pinia";
 import { computed, ref } from "vue";
 import { useRoute, useRouter } from "vue-router";
 import MessageTypes from "@/components/monitoring/messageTypes";
-import { useCookies } from "vue3-cookies";
 import * as MonitoringEndpoints from "../composables/serviceMonitoringEndpoints";
 import memoiseOne from "memoize-one";
 import { formatGraphDuration } from "../components/monitoring/formatGraph";
 import { useFailedMessageStore } from "./FailedMessageStore";
+import { useMonitoringHistoryPeriodStore } from "./MonitoringHistoryPeriodStore";
 
 export const useMonitoringStore = defineStore("MonitoringStore", () => {
   const failedMessageStore = useFailedMessageStore();
-  const { cookies } = useCookies();
-
-  const periods = [
-    { pVal: 1, text: "1m", refreshIntervalVal: 1 * 1000, refreshIntervalText: "Show data from the last minute. Refreshes every 1 second" },
-    { pVal: 5, text: "5m", refreshIntervalVal: 5 * 1000, refreshIntervalText: "Show data from the last 5 minutes. Refreshes every 5 seconds" },
-    { pVal: 10, text: "10m", refreshIntervalVal: 10 * 1000, refreshIntervalText: "Show data from the last 10 minutes. Refreshes every 10 seconds" },
-    { pVal: 15, text: "15m", refreshIntervalVal: 15 * 1000, refreshIntervalText: "Show data from the last 15 minutes. Refreshes every 15 seconds" },
-    { pVal: 30, text: "30m", refreshIntervalVal: 30 * 1000, refreshIntervalText: "Show data from the last 30 minutes. Refreshes every 30 seconds" },
-    { pVal: 60, text: "1h", refreshIntervalVal: 60 * 1000, refreshIntervalText: "Show data from the last hour. Refreshes every 1 minute" },
-  ];
-
-  function getHistoryPeriod(route = null, requestedPeriod = null) {
-    const period = requestedPeriod ?? (route?.query?.historyPeriod || cookies.get("history_period"));
-
-    return allPeriods.value.find((index) => index.pVal === parseInt(period)) ?? periods[0];
-  }
+  const historyPeriodStore = useMonitoringHistoryPeriodStore();
 
   function getPropertyValue(obj, path) {
     const properties = path.split(".");
@@ -54,7 +39,6 @@ export const useMonitoringStore = defineStore("MonitoringStore", () => {
     selectedGrouping: 0,
   });
 
-  const allPeriods = ref(periods);
   const endpointList = ref([]);
   const endpointName = ref("");
   const endpointDetails = ref({});
@@ -67,7 +51,6 @@ export const useMonitoringStore = defineStore("MonitoringStore", () => {
   const sortBy = ref("name");
   const isSortAscending = ref(false);
   const isInitialized = ref(false);
-  const historyPeriod = ref(getHistoryPeriod());
   const endpointListCount = computed(() => endpointList.value.length);
   const endpointListIsEmpty = computed(() => endpointListCount.value === 0);
   const endpointListIsGrouped = computed(() => grouping.value.selectedGrouping !== 0);
@@ -75,7 +58,7 @@ export const useMonitoringStore = defineStore("MonitoringStore", () => {
 
   //STORE ACTIONS
   async function initializeStore() {
-    await setHistoryPeriod();
+    //await setHistoryPeriod();
     await updateFilterString();
     await updateEndpointList();
     isInitialized.value = true;
@@ -96,7 +79,7 @@ export const useMonitoringStore = defineStore("MonitoringStore", () => {
   }
 
   async function updateEndpointList() {
-    endpointList.value = await MonitoringEndpoints.useGetAllMonitoredEndpoints(historyPeriod.value.pVal);
+    endpointList.value = await MonitoringEndpoints.useGetAllMonitoredEndpoints(historyPeriodStore.historyPeriod.pVal);
     if (!endpointListIsEmpty.value) {
       sortEndpointList();
       updateGroupSegments();
@@ -182,20 +165,6 @@ export const useMonitoringStore = defineStore("MonitoringStore", () => {
     disconnectedEndpointCount.value = await MonitoringEndpoints.useGetDisconnectedEndpointCount();
   }
 
-  /**
-   * @param {String} period - The history period value
-   * @description Sets the history period based on, in order of importance, a passed parameter, the url query string, saved cookie, or default value
-   */
-  async function setHistoryPeriod(requestedPeriod = null) {
-    const period = getHistoryPeriod(route, requestedPeriod);
-
-    if (period) {
-      historyPeriod.value = period;
-      cookies.set("history_period", historyPeriod.value.pVal);
-      await router.replace({ query: { ...route.query, historyPeriod: historyPeriod.value.pVal } });
-    }
-  }
-
   async function updateSort(newSortBy = "name", newIsSortAscending = false) {
     sortBy.value = newSortBy;
     isSortAscending.value = newIsSortAscending;
@@ -225,7 +194,6 @@ export const useMonitoringStore = defineStore("MonitoringStore", () => {
   return {
     //state
     grouping,
-    allPeriods,
     endpointList,
     endpointName,
     endpointDetails,
@@ -235,7 +203,6 @@ export const useMonitoringStore = defineStore("MonitoringStore", () => {
     disconnectedEndpointCount,
     negativeCriticalTimeIsPresent,
     filterString,
-    historyPeriod,
     sortBy,
     isSortAscending,
     isInitialized,
@@ -248,7 +215,6 @@ export const useMonitoringStore = defineStore("MonitoringStore", () => {
 
     //actions
     initializeStore,
-    setHistoryPeriod,
     updateSelectedGrouping,
     updateMessageTypes,
     updateSort,
